# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for what went wrong",
        "title": "What Went Wrong API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/me/plan": {
            "get": {
                "description": "Returns the user's current subscription plan and their active entitlements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plan"
                ],
                "summary": "Get current user plan and entitlements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetMePlanResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.PlanFetchErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Updates the user's subscription plan (e.g., from 'free' to 'premium').",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plan"
                ],
                "summary": "Update user plan",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PostMePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PostMePlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.PlanValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.PlanUpdateErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.GetMePlanResponse": {
            "type": "object",
            "properties": {
                "entitlements": {
                    "description": "Entitlements struct might need examples in its own definition if not here",
                    "allOf": [
                        {
                            "$ref": "#/definitions/services.Entitlements"
                        }
                    ]
                },
                "plan": {
                    "type": "string",
                    "example": "premium"
                }
            }
        },
...
        "handlers.PlanFetchErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "プランの取得に失敗しました"
                }
            }
        },
        "handlers.PlanUpdateErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "プランの更新に失敗しました"
                }
            }
        },
        "handlers.PlanValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "入力内容が正しくありません"
                }
            }
        },
        "handlers.PostMePlanRequest": {
            "type": "object",
            "required": [
                "plan"
            ],
            "properties": {
                "plan": {
                    "type": "string",
                    "example": "premium"
                }
            }
        },
        "handlers.PostMePlanResponse": {
            "type": "object",
            "properties": {
                "entitlements": {
                    "$ref": "#/definitions/services.Entitlements"
                },
                "plan": {
                    "type": "string",
                    "example": "premium"
                }
            }
        },
...
        "services.Entitlements": {
            "type": "object",
            "properties": {
                "canUseAiExcuse": {
                    "type": "boolean"
                },
                "canUsePremiumTemplates": {
                    "type": "boolean"
                },
                "logRetentionDays": {
                    "description": "nil = unlimited",
                    "type": "integer"
                },
                "maxGoals": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}